+++ .gitignore
@@ -1,0 +1,2 @@
+
+test/suites/chip-tool/chip-tool.log
