+++ test/action.yml
@@ -1,4 +1,4 @@
-name: EdgeX Snap Tester
+name: Matter Snap Tester
 description: |
   This Github action runs various smoke tests to validate the snap packaging.
 
@@ -17,16 +17,6 @@ inputs:
     description: |
       Channel for downloading the snap from store.
       This is useful only when 'snap' input is not set.
-    required: false
-  platform_channel:
-    description: |
-      Channel for downloading the edgexfoundry snap from store.
-      This is used when testing services against the edgexfoundry snap.
-      This is useful only when 'snap' input is not set.
-    required: false
-  full_config_test:
-    description: |
-      Option for enabling full config options test to include for e.g. legacy options testing.
     required: false
   go-args:
     description: |
@@ -60,9 +50,6 @@ runs:
     - shell: bash
       working-directory: ${{github.action_path}}
       env:
-        LOCAL_PLATFORM_SNAP: ${{ inputs.name == 'edgexfoundry' && steps.path.outputs.local_snap || '' }}
-        LOCAL_SERVICE_SNAP: ${{ inputs.name != 'edgexfoundry' && steps.path.outputs.local_snap || '' }}
-        PLATFORM_CHANNEL: ${{inputs.platform_channel}}
         SERVICE_CHANNEL: ${{inputs.channel}}
         FULL_CONFIG_TEST: ${{inputs.full_config_test}}
         SKIP_TEARDOWN_REMOVAL: ${{inputs.skip_teardown_removal}}
