--- test/suites/edgexfoundry/.gitignore
@@ -1,3 +1,0 @@
-ca.cert
-ca.key
-server.csr
