--- test/suites/edgexfoundry/main_test.go
@@ -1,87 +1,0 @@
-package test
-
-import (
-	"edgex-snap-testing/test/utils"
-	"log"
-	"os"
-	"testing"
-	"time"
-)
-
-const (
-	platformSnap            = "edgexfoundry"
-	supportSchedulerApp     = "support-scheduler"
-	supportSchedulerService = platformSnap + "." + supportSchedulerApp
-)
-
-func TestMain(m *testing.M) {
-	teardown, err := setup()
-	if err != nil {
-		log.Fatalf("Failed to setup tests: %s", err)
-	}
-
-	code := m.Run()
-	teardown()
-
-	os.Exit(code)
-}
-
-func TestCommon(t *testing.T) {
-
-	// The common config tests aren't used
-	// - TestAppConfig is covered in local startup message testing
-	// - TestGlobalConfig and TestMixedGlobalAppConfig can't be used because this
-	// snap multiple servers, test setting startup message instead
-	// - TestAutostart testing fails (unknown reason)
-	// utils.TestConfig(t, platformSnap, utils.Config{
-	// 	TestAutoStart: true,
-	// })
-
-	utils.TestNet(t, platformSnap, utils.Net{
-		StartSnap:        false, // the service are started by default
-		TestOpenPorts:    utils.PlatformPorts(true),
-		TestBindLoopback: utils.PlatformPorts(false), // exclude public ports
-	})
-
-	utils.TestPackaging(t, platformSnap, utils.Packaging{
-		TestSemanticSnapVersion: true,
-	})
-}
-
-func setup() (teardown func(), err error) {
-	log.Println("[CLEAN]")
-	utils.SnapRemove(nil, platformSnap)
-
-	log.Println("[SETUP]")
-	start := time.Now()
-
-	teardown = func() {
-		log.Println("[TEARDOWN]")
-		utils.SnapDumpLogs(nil, start, platformSnap)
-
-		log.Println("Removing installed snap:", !utils.SkipTeardownRemoval)
-		if !utils.SkipTeardownRemoval {
-			utils.SnapRemove(nil, platformSnap)
-		}
-	}
-
-	if utils.LocalPlatformSnap() {
-		err = utils.SnapInstallFromFile(nil, utils.LocalPlatformSnapPath)
-	} else {
-		err = utils.SnapInstallFromStore(nil, platformSnap, utils.PlatformChannel)
-	}
-	if err != nil {
-		teardown()
-		return
-	}
-
-	utils.SnapStart(nil, platformSnap)
-
-	// make sure all services are online before starting the tests
-	if err = utils.WaitPlatformOnline(nil); err != nil {
-		teardown()
-		return
-	}
-
-	return
-}
