--- test/suites/edgexfoundry/refresh/refresh_test.go
@@ -1,12 +1,0 @@
-package test
-
-import (
-	"edgex-snap-testing/test/utils"
-	"testing"
-)
-
-const platformSnap = "edgexfoundry"
-
-func TestCommon(t *testing.T) {
-	utils.TestRefresh(t, platformSnap)
-}
